Unsuccesfull child exits cause EIO return, handle the exception as EOF.
Move self.connected to be implicit, based on whether self.fd is None
or not. Add __getattr__() to provide the attribute.

setblocking() can be called when not connected, it will remember the
state and do as asked when connection is started. We don't really have
an fd until we're connected, as opposed to real sockets.

Be more careful invalidating self.fd, and checking that it is set.